Allow uuid to be specified.
Current deployment scripts perform db updates and then, config-import. For vocabularies created locally, that means the config-exported will have some given UUID, but the Vocabulary created in a remote environment via ce_deploy a different one. 

The result is that the follow-up config-import done by the deployment will crash, as it'll attempt to create the same vocabulary that already exists.

Specifying the 'uuid' in our .yml files prevents this.